# Conflicts:
#	src/main/java/seedu/address/ui/InventoryListPanel.java
#	src/main/java/seedu/address/ui/PersonCard.java
#	src/main/java/seedu/address/ui/PersonListPanel.java
#	src/main/resources/view/DrinkDetailPane.fxml

+++ src/main/java/seedu/address/model/drink/Price.java
@@ -41,7 +41,7 @@ public class Price {
      * Returns true if a given string is a valid price.
      */
     public static boolean isValidPrice(String test) {
-        return test.matches(PRICE_VALIDATION_REGEX) && Float.parseFloat(test) < Float.MAX_VALUE;
+        return test.matches(PRICE_VALIDATION_REGEX) && Double.parseDouble(test) < Float.MAX_VALUE;
     }
 
     @Override
